code changes

+++ src/main/java/Data/models/ProductPojo.java
@@ -3,10 +3,12 @@ package Data.models;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @Getter
 public class ProductPojo {
     private String productName;
-    private Double productPrice;
+    private BigDecimal productPrice;
     private Boolean inCart;
 }
